Update README.md
I've updated the layout of the badges. The second and third badge measure something differently but have the exact same text, making them less useful.

I didn't see an option to change the text inside the button, but by putting them in a table with headers, we now have some context as to what these buttons actually mean.